Add ability to spread attributes with `:attrs=`
A dictionary value can be "spread" into the set of attributes to
be provided to the component. This runs recursively so further
dynamic values can be provided through this method too.

Complex attributes cannot be provided through this spread, as
that would require the template rendering itself to become
recursive.